backend/requirements.txt
flask
flask-pymongo
requests
google-auth
google-auth-oauthlib
flask-cors
google.generativeai
python-dotenv
